backend/requirements.txt
asgiref==3.11.0
Django==5.2.8
sqlparse==0.5.3
typing_extensions==4.15.0
psycopg2-binary>=2.9
psycopg2-binary>=2.9
psycopg2-binary>=2.9
djangorestframework
django-cors-headers